move types definitions out of typecollector
